requirements.txt
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.37
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Cors==5.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
alembic==1.14.0
Werkzeug==3.1.3
cryptography==44.0.0
